Fix bad adapter registration causing problems in dexterity factory.
Traversing ++add++Document was resulting in the error:
AttributeError: 'MapWidgets' object has no attribute '__of__'

The problem is that within ++add++* an multi adapter lookup is made, where
the adapter is not named, has no specific provided interface and adapts
three objects.

collective.geo.mapwidget registered the adapter for MapWidgets without
a provided interface and with unspecific three-tuple discriminators,
causing to be falsely adapted in the factory code.

The fix is to simply be a bit more specific in the adapter registration.